requirements.txt
PyQt5==5.13.0
pyqt5-tools==5.15.1.2
PyQt5-stubs==5.14.2.2
pillow
numpy